:scissors: refactor(app): move List subcomponents to separate folder
