+++ app/src/components/list/Blip.vue
@@ -113,7 +113,7 @@
 import { Component, Vue, Prop } from 'vue-property-decorator'
 import copy from 'clipboard-copy'
 import MarkdownIt from 'markdown-it'
-import appConfig from '../config'
+import appConfig from '@/config'
 import { Blip as IBlip, Meta } from '@/types/domain'
 
 const md = new MarkdownIt()
