+++ app/src/components/list/NewBlip.vue
